Add RealmKeyedCollection swift file, docs
